requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
PyJWT==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
psycopg[binary]==3.2.9
gunicorn==23.0.0
flask-migrate==4.0.0
flask-login==0.6.2